requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioresponses==0.7.8
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
feedparser==6.0.11
filelock==3.18.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.2.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.0
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.24.1
setuptools==77.0.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
yfinance==0.2.55
